+++ tests/infer_missing_types_test/src/001_vaguer_comment_return.php
@@ -1,0 +1,12 @@
+<?php
+/** @return mixed */
+function test1() { return new ArrayObject(); }
+
+/** @return object */
+function test2() {
+  static $instance;
+  if ($instance === null) {
+    $instance = new ArrayObject();
+  }
+  return $instance;
+}